Test e2e working with cypress run
