+++ config.test.ci.yaml
@@ -23,7 +23,7 @@ services:
       cors: '*'
       debug: true
       recaptcha_secret: {env(RECAPTCHA_SECRET)}
-      recaptcha_sitekey: 6Lc2BtYUAAAAAOUBI-9r1sDJUIfG2nt6C43noOXh
+      recaptcha_sitekey: {env(RECAPTCHA_SITEKEY)}
       db:
         connection_uri: {env(CI_MONGODB_URI)}
         options:
